// frontend/src/app/shared/generated-api/model/auditEvent.ts
/**
 * Glacier Notes Cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AuditEvent { 
    id: string;
    eventType: string;
    occurredAt: string;
    actorUserId?: string;
    targetUserId?: string;
    targetEntityType?: string;
    targetEntityId?: string;
    result: AuditEventResultEnum;
    ipAddress?: string;
    clientDescription?: string;
    correlationId: string;
    metadata: { [key: string]: string; };
}
export enum AuditEventResultEnum {
    Success = 'SUCCESS',
    Failure = 'FAILURE',
    Denied = 'DENIED'
};